합치긱

+++ Back/readforce/src/main/java/com/readforce/entity/LiteratureQuiz.java
@@ -55,7 +55,7 @@ public class LiteratureQuiz {
 	private String explanation;
 	
 	@Column(nullable = false)
-	private Double score;
+	private int score;
 	
 	@CreatedDate
 	@Column(name = "created_date")
